+++ article.h
@@ -10,11 +10,11 @@ private:
     QString source;
     bool isScientificPaper;
     QString doi;
-    quint16 issueNumber;
+    int issueNumber;
 
 public:
     Article();
-    Article(const QString& coverImageUrl, const QString& title, const int publicationYear, const int uid, const QString& source, const QString& doi, quint16 issueNumber, bool isScientificPaper = false);
+    Article(const QString& coverImageUrl, const QString& title, const int publicationYear, const int uid, const QString& source, const QString& doi, int issueNumber, bool isScientificPaper = false);
     IMedia* clone() const override;
     QList<MediaParam> getMediaParams() const override;
     void accept(MediaVisitor& visitor) override {
